First commit.  TorrentParser, MetaInfo, and TorrentClient are the main objects that do the work.  I'm considering removing the TorrentParser and splitting its functionality between MetaInfo and main.py main().
The program is able to parse a torrent file, connect to a http tracker, and
parse the peer list that the tracker sends.  Unfortunately it can only do it
for one torrent file.

Sorry but in a bout of carelessness I forgot to test my assertions and it seems
like this commit WILL NOT RUN.  My next commit will work.